move Category interface to store/types, remove config, fix imports
