--- src/config.ts
@@ -1,5 +1,0 @@
-export interface Category {
-  board: string;
-  name: string;
-  description: string;
-}
